// src/enum/enums.ts
export enum Gender {
    male = 'male',
    female = 'female',
    other = 'other',
}

export enum Day {
    monday = 'monday',
    tuesday = 'tuesday',
    wednesday = 'wednesday',
    thursday = 'thursday',
    friday = 'friday',
    saturday = 'saturday',
    sunday = 'sunday',
}

export enum Role {
    admin = 'admin',
    teacher = 'teacher',
    student = 'student'
}